backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow>=10.4.0
numpy<2,>=1.24.3
opencv-python==4.8.1.78
pytesseract==0.3.10
anthropic==0.40.0
openai==1.30.0
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
